requirements.txt
# requirements.txt
Django==4.2.3
channels==4.0.0
celery==5.3.1
redis==4.6.0
djangorestframework==3.14.0
psycopg2-binary==2.9.6
daphne==4.0.0
channels-redis==4.1.0
django-cors-headers==4.2.0
Pillow==10.0.0  # For image processing